chore: move functions to test utils
